fixed installmentCount (float -> int), and removed it from encashment request

+++ src/app/shared/services/approvals/loan.service.ts
@@ -93,7 +93,7 @@ class LoanApprovalAdapter {
       loanCode: apiSchema.loanCode,
       loanName: apiSchema.loanName,
       totalAmount: parseFloat(apiSchema.totalAmount),
-      installmentCount: parseFloat(apiSchema.installmentCount),
+      installmentCount: parseInt(apiSchema.installmentCount),
       status: apiSchema.status,
       remarks: apiSchema.remarks,
     };
